Create `isPausedOrMaxTabsReached()` and `interruptTabOpening()`.
`isPausedOrMaxTabsReached()` is the condition that will replace simply `pauseState.isPaused`. It is slightly more complicated than it would otherwise need to be, in order to play nice with the `openTabsSameWindow` setting.

`interruptTabOpening()` adds the listeners to listen for the events that may resume tab opening, and checks `isPausedOrMaxTabsReached()` to see if resuming is indeed correct.